// client/src/components/brackets/types.tsx
export type ParticipantType = {
  id: string | number;

  isWinner?: boolean;

  name?: string;

  status?: "PLAYED" | "NO_SHOW" | "WALK_OVER" | "NO_PARTY" | string | null;

  resultText?: string | null;

  [key: string]: any;
};

export type MatchType = {
  id: number | string;

  /** Link to this match. While onClick() can be used, providing an href
      better supports opening a new tab, or copying a link. * */
  href?: string;

  name?: string;

  nextMatchId: number | string | null;

  swissRoundNumber?: number; // Used only for swiss to keep track of which column (rather than nextMatchId)

  nextLooserMatchId?: number | string | null;

  tournamentRoundText?: string;

  startTime: string | null;

  state: "PLAYED" | "NO_SHOW" | "WALK_OVER" | "NO_PARTY" | string;

  participants: ParticipantType[];

  isThirdPlaceMatch?: boolean;

  isExhibitionMatch?: boolean;

  [key: string]: any;
};
// ...
export type MatchComponentProps = {
  match: MatchType;

  onMatchClick: (args: {
    match: MatchType;
    topWon: boolean;
    bottomWon: boolean;
    event: React.MouseEvent<HTMLAnchorElement, MouseEvent>;
  }) => void;

  onPartyClick: (party: ParticipantType, partyWon: boolean) => void;

  onMouseEnter: (partyId: string | number) => void;

  onMouseLeave: () => void;

  topParty: ParticipantType;

  bottomParty: ParticipantType;

  topWon: boolean;

  bottomWon: boolean;

  topHovered: boolean;

  bottomHovered: boolean;

  topText: string;

  bottomText: string;

  connectorColor?: string;

  computedStyles?: ComputedOptionsType;

  teamNameFallback: string;

  resultFallback: (participant: ParticipantType) => string;
};
// ...
export type CommonTreeProps = {
  svgWrapper?: (props: {
    bracketWidth: number;
    bracketHeight: number;
    startAt: number[];
    children: ReactElement;
  }) => React.ReactElement;

  theme?: ThemeType;

  options?: { style: OptionsType };
};

export type BracketLeaderboardProps = CommonTreeProps & {
  matchComponent: (props: MatchComponentProps) => JSX.Element;

  currentRound?: string;

  onMatchClick?: (args: {
    match: MatchType;
    topWon: boolean;
    bottomWon: boolean;
  }) => void;

  onPartyClick?: (party: ParticipantType, partyWon: boolean) => void;
};

export type SwissLeaderboardProps = BracketLeaderboardProps & {
  matches: MatchType[];
};